requirements.txt
aiosignal==1.3.2
albucore==0.0.19
albumentations==1.4.20
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.3.0
blosc2==2.7.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
et_xmlfile==2.0.0
eval_type_backport==0.2.0
executing==2.1.0
filelock==3.16.1
fire==0.7.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
gmpy2==2.2.1
idna==3.10
imageio==2.36.0
intel-cmplr-lib-ur==2025.0.4
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
knit==0.2.4
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mlai==0.1.2
mpmath==1.3.0
msgpack==1.1.0
ndindex==1.9.2
nest-asyncio==1.6.0
networkx==3.4.2
notutils==0.0.1
numexpr==2.10.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.0.0
platformdirs==4.3.6
pods==0.1.14
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==18.1.0
pybboxes==0.1.6
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.40.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sahi==0.11.20
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
simsimd==6.2.1
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.4
stringzilla==3.11.1
sympy==1.13.1
tables==3.10.1
tbb==2022.0.0
tcmlib==1.2.0
tensorboardX==2.6.2.2
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.55
ultralytics-thop==2.0.13
umf==0.9.1
urllib3==2.2.3
wcwidth==0.2.13
